refactor: extract vocab summary from page
